requirements.txt
pytest
pillow
numpy

